requirements.txt
# daTEAbase Requirements
# Python 依賴套件清單

# PostgreSQL 資料庫驅動
psycopg2-binary>=2.9.0

# MongoDB 驅動（用於數據分析，選用）
pymongo>=4.0.0

# 密碼加密
bcrypt>=4.0.0

# 日期時間處理
python-dateutil>=2.8.0

# 其他工具
colorama>=0.4.6  # 終端機顏色輸出（選用）
